% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sqf_exp}
\alias{sqf_exp}
\alias{sqf_unif}
\alias{sqf_norm}
\alias{sqf_cauchy}
\alias{sqf_halftriang}
\alias{sqf_halfcosine}
\alias{sqf_sech}
\title{Basic quantile functions}
\usage{
sqf_exp(u, ...)

sqf_unif(u, ...)

sqf_norm(u, ...)

sqf_cauchy(u, ...)

sqf_halftriang(u, ...)

sqf_halfcosine(u, ...)

sqf_sech(u, ...)
}
\arguments{
\item{u}{depth}

\item{...}{not used}
}
\value{
modified function
}
\description{
Basic quantile functions available for modification
\code{sqf_exp()}: Basic QF of exponential distribution
\code{sqf_unif()}: Basic QF of uniform distribution
\code{sqf_norm()}: Basic QF of normal distribution \code{qnorm()}.
\code{sqf_cauchy()}: Basic QF of Cauchy distribution\code{. }sqf_halftriang()\verb{: Basic QF of half-triangular distribution. }sqf_halfcosine()\verb{: Basic QF of half-cosine distribution. }sqf_sech()`: Basic QF of hyperbolic secant distribution.
}
\examples{
sqf_exp(0.6)
qexp(0.6, 1)
}
