% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_qf}
\alias{as_qf}
\alias{expects}
\alias{returns}
\alias{math}
\title{Wrapper for producing classed functions.}
\usage{
as_qf(fun, subclass = NULL, expects = NA, returns = NA, math = NA)

expects(fun)

returns(fun)

math(fun)
}
\arguments{
\item{fun}{function}

\item{subclass}{additional classes to be added to QF}

\item{expects}{domain flag for incoming function domain}

\item{returns}{domain flag for resulting function domain}

\item{math}{raw string for math describing quantile function}
}
\description{
Returns function of classes \code{c("function", "qf")}
Subclass can be added through optional \code{subclass} argument
}
