% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawmath.R
\name{vrbl}
\alias{vrbl}
\alias{br}
\alias{fn_insert}
\alias{prmtr}
\alias{display}
\alias{inline}
\title{Show raw math}
\usage{
vrbl(x)

br(x, left = "[", right = "]")

fn_insert(x, y, br = TRUE, left = "[", right = "]", pl = "&")

prmtr(x, .invert = FALSE)

display(.qf, prfx = "Q(u)=", pl = "&", u = "u")

inline(.qf, prfx = "Q(u)=", pl = "&", u = "u")
}
\arguments{
\item{x}{raw string. Math string to be displayed}

\item{left}{raw string. Left bracket}

\item{right}{raw string. Right bracket}

\item{y}{raw string. Math string to be inserted.}

\item{br}{logical. Should the inserted string be bracketed? Default is TRUE.}

\item{pl}{character. Single character inidicating the placeholder.}

\item{.invert}{logical. Should the parameter be inverted (1/x). Default is FALSE}

\item{prfx}{raw string. Prefix string. Default is Q(u)=}

\item{u}{character. Variable name, which will be placed instead of placeholder.}
}
\description{
Show raw math
}
\keyword{internal}
