% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{sqf_exp}
\alias{sqf_exp}
\alias{sqf_unif}
\title{Basic quantile functions}
\usage{
sqf_exp(u, ...)

sqf_unif(u, ...)
}
\arguments{
\item{u}{depth}

\item{...}{not used}
}
\value{
modified function
}
\description{
Basic quantile functions
}
\examples{
sqf_exp(0.6)
qexp(0.6, 1)
}
