% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{s_exp}
\alias{s_exp}
\alias{s_unif}
\alias{s_norm}
\alias{s_cauchy}
\alias{s_halftriang}
\alias{s_halfcosine}
\alias{s_sech}
\title{Basic quantile functions}
\usage{
s_exp(u, ...)

s_unif(u, ...)

s_norm(u, ...)

s_cauchy(u, ...)

s_halftriang(u, ...)

s_halfcosine(u, ...)

s_sech(u, ...)
}
\arguments{
\item{u}{depth}

\item{...}{not used}
}
\value{
modified function
}
\description{
Basic quantile functions available for modification
\code{s_exp()}: Basic QF of exponential distribution
\code{s_unif()}: Basic QF of uniform distribution
\code{s_norm()}: Basic QF of normal distribution \code{qnorm()}.
\code{s_cauchy()}: Basic QF of Cauchy distribution\code{. }s_halftriang()\verb{: Basic QF of half-triangular distribution. }s_halfcosine()\verb{: Basic QF of half-cosine distribution. }s_sech()`: Basic QF of hyperbolic secant distribution.
}
\examples{
s_exp(0.6)
qexp(0.6, 1)
}
