% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{qtr_decorate}
\alias{qtr_decorate}
\title{Decorate the basic QF with location and scale parameters}
\usage{
qtr_decorate(
  fun,
  nm_location = ".location",
  nm_scale = ".scale",
  location = 0,
  scale = 1,
  .invert = FALSE
)
}
\arguments{
\item{fun}{function to be decorated}

\item{nm_location}{character. The name of the location parameter. Default name \code{.location}. Default value is 0.}

\item{nm_scale}{character. The name of the scale parameter. Default name \code{.scale}. Default value is 1.}

\item{location}{numeric. Fixed value for location. Default is 0.}

\item{scale}{numeric. Fixed value for scale. Default is 1.}

\item{.invert}{logical. Should the scale parameter be inverted (1/.scale) before applying. Default FALSE}
}
\value{
modified function with location and scale parameter. The parameter names specified by the user.
}
\description{
Add location (\code{.location}) and scale (\code{.scale}) parameters (or the inverse scale for \code{qtr_idecorate()})
Note! The parameter names can be changed by passing the names
}
\examples{
qtr_decorate(s_exp)
}
