% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{ptr_power}
\alias{ptr_power}
\alias{ptr_exponentiate}
\alias{ptr_fun}
\alias{ptr_shiftby}
\alias{ptr_scaleby}
\alias{ptr_powerby}
\title{p-transformations}
\usage{
ptr_power(fun, nm_pow = ".pow", .invert = FALSE)

ptr_exponentiate(fun, nm_base = ".base", .invert = FALSE)

ptr_fun(fun, .fun)

ptr_shiftby(fun, x = 0)

ptr_scaleby(fun, x = 1)

ptr_powerby(fun, x = 1)
}
\arguments{
\item{fun}{function}

\item{nm_pow}{character.  The name of the power parameter. The default name is \code{.pow}. The default value is 1
Should be a valid unique variable name other than "u"}

\item{.invert}{logical. Should the power parameter be inverted (1/.pow) before applying. Default FALSE}

\item{nm_base}{character.  The name of the base parameter. The default name is \code{.base}. The default value is \code{exp(1)} (Euler's constant).}

\item{.fun}{function without arguments(or with all default arguments) to be applied as Q-transformation}

\item{x}{numeric. Fixed value to shift/scale/power the u by}
}
\value{
modified function
}
\description{
Some of the typical transformations of QFs, implementing a Q-transformation rule.
\code{ptr_power()}: Raising of QF to a power. Returns \eqn{Q_1(u)^k}.
\code{ptr_ipower()}: Raising of QF to an inverse power. Returns \eqn{Q_1(u)^{1/k}}.
\code{ptr_exponentiate()}: Exponentiating the QF. Returns \eqn{k^Q_1(u)}.
\code{ptr_fun()}: Q-transform with generic function without additional arguments. \eqn{.fun(Q_1(u))}.
}
\examples{
qf_exp <- function(u)-log(1-u)
qf_weibull <- qtr_power(qf_exp, "k")
qf_weibull(0.5,k=1/5)
qweibull(0.5, shape = 5)
qtr_fun(sqf_exp,log1p)
}
