% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{qff_reflect}
\alias{qff_reflect}
\alias{qff_add}
\alias{qff_multiply}
\alias{qff_mix}
\alias{qff_imix}
\alias{qff_reciprocate}
\alias{qff_decorate}
\title{Gilchrist rules}
\usage{
qff_reflect(fun)

qff_add(fun1, fun2)

qff_multiply(fun1, fun2)

qff_mix(fun1, fun2)

qff_imix(fun1, fun2)

qff_reciprocate(fun)

qff_decorate(fun)
}
\arguments{
\item{fun, fun1, fun2}{functions}
}
\value{
modified function
}
\description{
Gilchrist rules
}
\examples{
qf_exp <- function(u)-log(1-u)
qf_logistic <- qff_add(qf_exp, qff_reflect(qf_exp))
qf_logistic(0.6)
qlogis(0.6)
}
